feat: Add new dashboard components for exams, announcements, assignments, grades, and resources refactor: Update class enrollments and schedule models for consistency

+++ app/Models/Schedule.php
@@ -71,7 +71,7 @@ final class Schedule extends Model
     /**
      * Get the duration of the class in minutes
      */
-    public function getDurationMinutesAttribute(): int
+    public function getDurationMinutesAttribute(): float
     {
         if (! $this->start_time || ! $this->end_time) {
             return 0;
